pyproject.toml
[project]
name = "countloc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.43",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
